Clean up more data utilities
