--- tests/data/fake_psfs.dat
@@ -1,3 +1,0 @@
-# visit_id psf_val
-000002 1.3
-000012 1.5
