--- tests/data/fake_times.dat
@@ -1,4 +1,0 @@
-# visit_id mean_julian_date
-000003 57162.0
-000005 57172.0
-010006 100000.0
