--- tests/data/mpcs.txt
@@ -1,3 +1,0 @@
-     Fake1    C2001 05 05.03500 18 45 47.64 -24 27 20.3          16.6 R      706
-     Fake2   2C1995 01 20.70000 23 29 42.54 -02 59 15.0          15.5 V      706
-     Fake3    C2005 10 10.28000 01 45 13.20 +08 07 21.4                r     706
